requirements.txt
dnspython
requests
marrow.mailer
texttable
